First Prototype
This is only simulating the exchange of minerals and water between the pot and the plant and the growth of leaves and roots.

Every plant will have an excess and deficiency level for each mineral and water. Absorption strength > 0 means the plant will actively transport that mineral.

Growing leaves or roots or active transport takes energy and eats up plant resources.

Tick period on the pot class is how often the exchange happens. growTick in plant class is how often the plant checks to grow.

Currently there are no consequences for something to be in an inappropriate amount.