Prototype Complete
Open the scene in Assets/Umer folder.
The pot object has a mud object child which has a "Soil" script attached to it.
In it there is a list of 5 elements and the top one is water.
If u keep watering the plant (Entering 5 in the water index) the plant will continue to grow. Only water the plant if the levels get low.